refactor: move Variable class into new variable.py
