+++ gmatpyplus/parameter/__init__.py
@@ -1,1 +1,2 @@
 from .parameter import *
+from .variable import Variable
